src/test/javascript/spec/app/entities/google-metric-group/google-metric-group.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { GoogleMetricGroupService } from 'app/entities/google-metric-group/google-metric-group.service';
import { IGoogleMetricGroup, GoogleMetricGroup } from 'app/shared/model/google-metric-group.model';

describe('Service Tests', () => {
  describe('GoogleMetricGroup Service', () => {
    let injector: TestBed;
    let service: GoogleMetricGroupService;
    let httpMock: HttpTestingController;
    let elemDefault: IGoogleMetricGroup;
    let expectedResult: IGoogleMetricGroup | IGoogleMetricGroup[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(GoogleMetricGroupService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new GoogleMetricGroup('ID', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find('123').subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a GoogleMetricGroup', () => {
        const returnedFromService = Object.assign(
          {
            id: 'ID'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new GoogleMetricGroup()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a GoogleMetricGroup', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            googleId: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of GoogleMetricGroup', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            googleId: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a GoogleMetricGroup', () => {
        service.delete('123').subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
